Update universal.js
Combined onresize trigger to trigger both closeSide and scrollProgress at the same time, rather than one after the other.

Completely overhauled scrollProgress function, heavily simplifying it and removing a bunch of unnecessary checks and variables.

Removed scrollProgress trigger in toggleSide, as the new format for the progress bar no longer requires moving it aside.

Removed unused logs in the closeSide function, and also heavily simplified it. It now just calls the toggleSide function if it sees that the sidemenu is still open while the sidebutton is no longer active.