Main body updates
Import google font Mate for use as the main font, since it works well alongside the courier font used for the logo.

implement content-box div, which is used to position the content in the center of the screen. By default its a column flex, but if the screen is wide enough, it converts to a row flex.

Make it so the text and image of the homepage are divided evenly in the content box.